DEVX-2289: Simplifies configuration in orders-service (#51)
This sets the stage for configuring entirely using Spring.  Right now a helper
script is required to preprocess some configurations.  However, using
active contexts w/ Spring, we should be able to load multiple configuration
files and use the ability to replace one config with value of another to remove
the preprocessing script from the application image